update changelog and redirect to home on click when user is not logged in

+++ pages/app/embed/EmbedWidget.js
@@ -11,7 +11,7 @@ import { initStore } from 'store';
 import { bindActionCreators } from 'redux';
 import { getWidget } from 'redactions/widget';
 import { setUser } from 'redactions/user';
-import { setRouter } from 'redactions/routes';
+import { setRouter, Router } from 'redactions/routes';
 
 // Components
 import Page from 'components/app/layout/Page';
@@ -129,26 +129,35 @@ class EmbedWidget extends Page {
     const { favorite } = this.state;
     const { widget, user } = this.props;
 
-    this.setState({ isLoading: true });
+    if (user.id) {
+      this.setState({ isLoading: true });
 
-    if (favorite) {
-      this.userService.deleteFavourite(favorite.id, user.token)
-        .then(() => {
-          this.setState({
-            favorite: null,
-            isLoading: false
-          });
-        })
-        .catch(err => toastr.error('Error unfavoriting the widget', err));
+      if (favorite) {
+        this.userService.deleteFavourite(favorite.id, user.token)
+          .then(() => {
+            this.setState({
+              favorite: null,
+              isLoading: false
+            });
+          })
+          .catch(err => toastr.error('Error unfavoriting the widget', err));
+      } else {
+        this.userService.createFavouriteWidget(widget.id, user.token)
+          .then((res) => {
+            this.setState({
+              favorite: res.data,
+              isLoading: false
+            });
+          })
+          .catch(err => toastr.error('Error setting the widget as favorite', err));
+      }
     } else {
-      this.userService.createFavouriteWidget(widget.id, user.token)
-        .then((res) => {
-          this.setState({
-            favorite: res.data,
-            isLoading: false
-          });
-        })
-        .catch(err => toastr.error('Error setting the widget as favorite', err));
+      toastr.confirm('You need to be logged in in order to favorite a widget. Click OK to sign up to RW!', {
+        onOk: () => {
+          // TODO We need to redirect to the sign up page instead of to the home page
+          Router.pushRouter('/');
+        }
+      });
     }
   }
